refactor project로부터 Initiative관련 코드 제거
