+++ src/main/java/kr/jay/okrver3/application/project/ProjectFacade.java
@@ -15,7 +15,6 @@ import kr.jay.okrver3.domain.project.service.ProjectService;
 import kr.jay.okrver3.domain.project.service.ProjectTeamMemberInfo;
 import kr.jay.okrver3.domain.user.User;
 import kr.jay.okrver3.domain.user.service.UserService;
-import kr.jay.okrver3.interfaces.project.ProjectDetailRetrieveCommand;
 import kr.jay.okrver3.interfaces.project.ProjectKeyResultSaveDto;
 import kr.jay.okrver3.interfaces.project.ProjectMasterSaveDto;
 import kr.jay.okrver3.interfaces.project.ProjectSideMenuResponse;
@@ -98,7 +97,4 @@ public class ProjectFacade {
 		return projectService.registerKeyResult(projectKeyResultSaveDto, user);
 	}
 
-	public String registerInitiative(ProjectInitiativeSaveCommand command, User user) {
-		return projectService.registerInitiative(command, user);
-	}
 }
