+++ src/main/java/kr/jay/okrver3/domain/project/service/ProjectRepository.java
@@ -4,9 +4,9 @@ import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 
+import kr.jay.okrver3.application.project.ProjectDetailRetrieveCommand;
 import kr.jay.okrver3.domain.project.Project;
 import kr.jay.okrver3.domain.user.User;
-import kr.jay.okrver3.interfaces.project.ProjectDetailRetrieveCommand;
 
 public interface ProjectRepository {
 	Project save(Project project);
