+++ src/main/java/kr/jay/okrver3/domain/project/service/ProjectService.java
@@ -4,9 +4,8 @@ import java.util.List;
 
 import org.springframework.data.domain.Page;
 
-import kr.jay.okrver3.application.project.ProjectInitiativeSaveCommand;
+import kr.jay.okrver3.application.project.ProjectDetailRetrieveCommand;
 import kr.jay.okrver3.domain.user.User;
-import kr.jay.okrver3.interfaces.project.ProjectDetailRetrieveCommand;
 import kr.jay.okrver3.interfaces.project.ProjectKeyResultSaveDto;
 import kr.jay.okrver3.interfaces.project.ProjectMasterSaveDto;
 import kr.jay.okrver3.interfaces.project.ProjectSideMenuResponse;
@@ -25,8 +24,6 @@ public interface ProjectService {
 	ProjectSideMenuResponse getProjectSideMenuDetails(String projectToken, User user);
 
 	String registerKeyResult(ProjectKeyResultSaveDto projectKeyResultSaveDto, User user);
-
-	String registerInitiative(ProjectInitiativeSaveCommand command, User user);
 
 }
