+++ src/test/java/kr/jay/okrver3/application/project/ProjectFacadeTest.java
@@ -3,7 +3,6 @@ package kr.jay.okrver3.application.project;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
@@ -41,7 +40,6 @@ import kr.jay.okrver3.domain.user.service.impl.UserServiceImpl;
 import kr.jay.okrver3.infrastructure.notification.NotificationJDBCRepository;
 import kr.jay.okrver3.infrastructure.project.ProjectQueryDslRepository;
 import kr.jay.okrver3.infrastructure.project.ProjectRepositoryImpl;
-import kr.jay.okrver3.interfaces.project.ProjectDetailRetrieveCommand;
 import kr.jay.okrver3.interfaces.project.ProjectKeyResultSaveDto;
 import kr.jay.okrver3.interfaces.project.ProjectMasterSaveDto;
 import kr.jay.okrver3.interfaces.project.ProjectSideMenuResponse;
@@ -312,7 +310,7 @@ class ProjectFacadeTest {
 
 		assertThatThrownBy(() -> sut.registerKeyResult(new ProjectKeyResultSaveDto(projectToken, keyResultName), user))
 			.isInstanceOf(OkrApplicationException.class)
- 			.hasMessage(ErrorCode.USER_IS_NOT_LEADER.getMessage());
+			.hasMessage(ErrorCode.USER_IS_NOT_LEADER.getMessage());
 	}
 
 	@Test
@@ -330,28 +328,5 @@ class ProjectFacadeTest {
 			.hasMessage(ErrorCode.KEYRESULT_LIMIT_EXCEED.getMessage());
 
 	}
-
-	@Test
-	@Sql("classpath:insert-project-date.sql")
-	void 행동전략_추가시_기대하는_응답을_리턴한다_initiativeToken() throws Exception {
-
-		ProjectInitiativeSaveCommand requestDto = new ProjectInitiativeSaveCommand(
-			"key_wV6MX15WQ3DTzQMs",
-			"행동전략",
-			LocalDate.now().minusDays(10),
-			LocalDate.now().plusDays(10),
-			"행동전략 상세내용"
-		);
-
-		User user = em.createQuery("select u from User u where u.id = :userSeq", User.class)
-			.setParameter("userSeq", 3L)
-			.getSingleResult();
-
-		String response = sut.registerInitiative(requestDto, user);
-
-		assertThat(response).containsPattern(
-			Pattern.compile("initiative-[a-zA-Z0-9]{10}"));
-	}
-
 
 }
